+++ python/sglang/srt/layers/moe/ep_moe/layer.py
@@ -36,21 +36,19 @@ from sglang.srt.layers.quantization.base_config import (
     QuantizeMethodBase,
 )
 from sglang.srt.layers.quantization.fp8 import Fp8Config, Fp8MoEMethod
+from sglang.srt.layers.quantization.fp8_kernel import (
+    scaled_fp8_quant as sgl_scaled_fp8_quant,
+)
 from sglang.srt.model_executor.forward_batch_info import ForwardMode
 from sglang.srt.utils import DeepEPMode, is_cuda, is_hip, set_weight_attrs
 
 _is_cuda = is_cuda()
+_is_hip = is_hip()
 
-if _is_cuda:
-    from sglang.srt.custom_op import scaled_fp8_quant as sgl_scaled_fp8_quant
-else:
+if _is_hip:
     from vllm import _custom_ops as vllm_ops
 
 logger = logging.getLogger(__name__)
-
-_is_hip = is_hip()
-
-_buffer = None
 
 
 class GroupedGemmRunner(torch.nn.Module):
